backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
SQLAlchemy==2.0.23
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML Dependencies
groq==0.4.1

# File Processing
PyPDF2==3.0.1
Pillow==10.1.0
pytesseract==0.3.10

# Data Processing
pandas==2.1.3
numpy==1.25.2

# Monitoring & Observability
evidently==0.4.11

# Development & Testing
pytest==7.4.3
httpx==0.25.2

# Utilities
matplotlib==3.8.2
